// frontend/src/api/generated/models/customAbbreviationDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export interface CustomAbbreviationDto {
  id?: string;
  coachUserId?: string;
  /** @nullable */
  abbreviation?: string | null;
  /** @nullable */
  fullText?: string | null;
  /** @nullable */
  description?: string | null;
}